feat: do not report events when the ticket is moved up or down the list
