+++ src/commands/start_webhook.rs
@@ -11,62 +11,12 @@ use axum::{
 use base64::{engine::general_purpose::STANDARD as Base64, Engine};
 use hmac::{Hmac, Mac};
 use hyper::{body::Bytes, HeaderMap};
-use reqwest::{Client, StatusCode};
+use reqwest::StatusCode;
 use sha1::Sha1;
 use tower_http::trace::{DefaultOnResponse, TraceLayer};
 use tracing::{debug, warn, Level};
 
-use crate::{
-    config::Config,
-    models::{DiscordEmbed, DiscordEmbedAuthor, DiscordWebhookEvent, MemberCreator, WebhookEvent},
-};
-
-struct DiscordReporter {
-    endpoint: String,
-    client: Client,
-}
-
-impl DiscordReporter {
-    pub fn new(endpoint: String) -> Self {
-        Self {
-            endpoint,
-            client: Client::new(),
-        }
-    }
-
-    pub async fn report(&self, event: WebhookEvent) -> Result<()> {
-        let mut card_name: Option<String> = None;
-        let mut card_url: Option<String> = None;
-        if let Some(card) = event.action.data.card {
-            card_name = Some(card.name);
-            card_url = Some(format!("https://trello.com/c/{}", card.short_link));
-        }
-        let discord_event = DiscordWebhookEvent {
-            embeds: vec![DiscordEmbed {
-                title: Some(event.action._type),
-                _type: Some(String::from("rich")),
-                description: card_name,
-                url: card_url,
-                color: Some(0x00ff00),
-                fields: vec![],
-                timestamp: Some(event.action.date),
-                author: Self::get_author(event.action.member_creator),
-            }],
-        };
-        self.client
-            .post(&self.endpoint)
-            .json(&discord_event)
-            .send()
-            .await?;
-        Ok(())
-    }
-
-    fn get_author(creator: Option<MemberCreator>) -> Option<DiscordEmbedAuthor> {
-        creator.map(|creator| DiscordEmbedAuthor {
-            name: creator.full_name.unwrap_or(creator.username),
-        })
-    }
-}
+use crate::{config::Config, models::WebhookEvent, reporting::DiscordReporter};
 
 struct WebhookState {
     pub reporter: DiscordReporter,
@@ -107,10 +57,7 @@ async fn post_endpoint(
         warn!("Could not parse payload: {}", e);
         (StatusCode::BAD_REQUEST, "Could not parse payload").into_response()
     })?;
-    debug!(
-        "New event: {} (date: {})",
-        event.action._type, event.action.date
-    );
+    debug!("New event: {:?}", raw_body);
     state.reporter.report(event).await.map_err(|e| {
         warn!("Internal server error: {}", e);
         (StatusCode::INTERNAL_SERVER_ERROR, "Internal server error").into_response()
